* commit '25ce90069e7ea9e096a7c3f56ae04dd64de3eb47': (75 commits)
  Updated with a Zenodo citation.
  Update README.md
  Minor tweak on framework search paths to get this to compile on iOS as a standalone.
  Updated documentation to reflect new framework targets.
  Added iOS module support for Xcode 6, so this can be used in a Swift application.
  Fixed some Xcode 6 compilation warnings / errors. Mac framework can now build as a module in Xcode 6.
  Removed the synchronous lock around movie processing. I see no reason this needs to be here, and it prevents multiple movies from being blended.
  Fixed a compilation error due to a missing property attribute.
  Fixed a bug causing movie recording crashes introduced by a pull request. Experimenting with improving recording performance using a separate movie recording OpenGL ES context.
  Fixed a setInputFrameBuffer issue in GPUImageThreeInputFilter
  Fix setFloatArray:... so that it copies the data in case it's on the stack
  Fixed a 64-bit crash with a Gaussian blur of 0 pixels.
  Fixed a crash on a 0-radius box blur.
  By default, now send an indefinite frame time with GPUImageUIElement updates to prevent freezing of the output. -updateUsingCurrentTime still lets you record movies from UI elements.
  fix GPUImageRawDataOutput memory access issues when working with CVPixelBuffer-backed framebuffer
  Made the pipeline additions a little more generic, so they wouldn't break on the Mac.
  Change filter input and outputs to the base class
  Add method to remove filter using only the filter
  Movie processing made more stable
  `dispatch_release` excluded on 6.0+ (instead of 7.0+)
  ...

Conflicts:
	framework/GPUImage.xcodeproj/project.pbxproj

+++ examples/iOS/BenchmarkSuite/BenchmarkSuite/VideoFilteringDisplayController.m
@@ -237,8 +237,8 @@
 
         unsigned char *data = (unsigned char *)CVPixelBufferGetBaseAddress(pixelBuffer);
 
-        int bufferHeight = CVPixelBufferGetHeight(pixelBuffer);
-        int bufferWidth = CVPixelBufferGetWidth(pixelBuffer);
+        size_t bufferHeight = CVPixelBufferGetHeight(pixelBuffer);
+        size_t bufferWidth = CVPixelBufferGetWidth(pixelBuffer);
         NSInteger myDataLength = bufferWidth * bufferHeight * 4;
